backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic-settings
python-jose
passlib[bcrypt]
python-multipart
celery
redis
